fix[vla]: harden pi05 security, correctness, consistency + widen quant thresholds
Security:
- fseeko with _FILE_OFFSET_BITS=64 for >2GB GGUF on 32-bit targets
- Sanity-check all GGUF hparams to prevent OOM from crafted files
- Null-check actions_out/n_actions_out in pi05_inference
- Validate n_heads/hidden before division in SigLIP block builder
- Validate LATEST_DATE format in mobile CI workflow

Correctness:
- Replace identity CHW permutation loop with direct pointer pass
- Assert vlm_n_layers == expert_n_layers at load time
- Read VLM hparams from GGUF instead of relying on hardcoded defaults
- Fix free order in model_factory.cpp (gguf_free before ggml_free)
- Remove dead iOS skip guards in test bodies
- Remove unnecessary ggml_cont on already-contiguous cached K/V

Performance:
- Flat K/V cache buffers (2 allocations vs 36)
- Vision tower weights constructed once outside camera loop
- Converter defers float32 cast to per-tensor encoding (halves peak RAM)

Consistency:
- Update stale Phase-1 stub comments to reflect Phase-3 implementation
- Fix llvm-cov-19 → unversioned llvm-cov in package.json
- Make prefill_compute_ms/prefill_total_ms required in TypeScript types
- Remove duplicate 0.1.0 changelog section
- Add weightsLoaded/destroyed to VlaModel state + _packageName/_packageVersion
- Rename CTest suite SmolvlaUnit → VlaGgmlUnit, raise timeout to 900s

Test thresholds:
- Widen M3.2–M3.6 parity bars to accommodate q_aggressive quantization
  noise (Q5_0 vision, Q8_0 VLM) and cross-hardware SIMD variance.
  Bars give ~2× headroom beyond worst observed values while still
  catching real regressions.

All 30 C++ tests pass locally with q_aggressive GGUF.
All 5 JS integration tests pass (62/62 assertions, cos≥0.999).

+++ packages/vla-ggml/scripts/convert_pi05_to_gguf.py
@@ -415,8 +415,9 @@ def load_state_dict_from_lerobot(checkpoint: str) -> dict[str, torch.Tensor]:
     policy = PI05Policy.from_pretrained(checkpoint, config=cfg, strict=False)
     policy.eval()
     sd = policy.state_dict()
-    # Detach + cast to float32 on CPU for deterministic downstream encoding.
-    return {k: v.detach().to(torch.float32).cpu() for k, v in sd.items()}
+    # Detach + move to CPU; keep original dtype to avoid 2× peak RAM.
+    # The per-tensor encode_for_gguf path handles the float32 cast.
+    return {k: v.detach().cpu() for k, v in sd.items()}
 
 
 def load_state_dict_from_safetensors(path: Path) -> dict[str, torch.Tensor]:
@@ -424,7 +425,7 @@ def load_state_dict_from_safetensors(path: Path) -> dict[str, torch.Tensor]:
     from safetensors.torch import load_file as st_load_file
 
     sd = st_load_file(str(path))
-    return {k: v.to(torch.float32) for k, v in sd.items()}
+    return {k: v for k, v in sd.items()}
 
 
 # ---------------------------------------------------------------------------
@@ -507,7 +508,7 @@ def convert(
             continue
 
         seen_pt_keys.add(pt_name)
-        arr = t.numpy()
+        arr = t.to(torch.float32).numpy()
         target = select_target_dtype(gg_name, arr, variant=variant)
         payload, raw_dtype = encode_for_gguf(arr, target)
         # For quantised tensors ``payload`` is a packed uint8 buffer whose
